rename Window to Classic, create superclass named Window.java, etc.
