+++ src/controller/ActionList.java
@@ -4,21 +4,21 @@ import javax.swing.JComboBox;
 
 import dao.ProgressDAO;
 import window.ManageBooks;
-import window.TestUI;
+import window.Window;
 
 public class ActionList {
-	private TestUI testUI;
+	private Window window;
 	private ManageBooks mBooks;
 	private ProgressDAO pdao;
 	private Controller controller;
 	
-	public ActionList(TestUI testUI) {
-		this.testUI = testUI;
+	public ActionList(Window window) {
+		this.window = window;
 	}
 
-	public void bookListButtonAction(int userId) {
+	public void bookListButtonAction(int userId, Window window) {
 		if (mBooks == null) {
-			mBooks = new ManageBooks(userId);
+			mBooks = new ManageBooks(userId, window);
 		}
 		mBooks.run();
 		mBooks.updateFrame(userId);
@@ -28,11 +28,14 @@ public class ActionList {
 		pdao = new ProgressDAO();
 		controller = new Controller();
 		String bookTitle;
-		JComboBox<Object> bookShelfCombo = new JComboBox<>(controller.setBookList(userId, bookId));
+		JComboBox<String> bookShelfCombo = new JComboBox<>(controller.setBookList(userId));
 		bookTitle = String.valueOf(bookShelfCombo.getSelectedItem());
 		bookId = pdao.searchBookId(userId, bookTitle);
-		testUI.updateText(userId, bookId);
+		
 	}
 
+//	public void reloadBookShelfCombo(int userId) {
+//		window.bookShelfCombo.setModel(controller.setBookList(userId));
+//	}
 
 }
