+++ src/window/TestUI.java
@@ -29,9 +29,8 @@ import controller.ActionList;
 import controller.Controller;
 import controller.ShowController;
 
-public class TestUI  {
+public class TestUI extends Window {
 
-	private static final long serialVersionUID = 1L;
 	private JPanel contentPane;
 	private JFrame frame;
 	protected int userId;
@@ -51,21 +50,22 @@ public class TestUI  {
 	private JProgressBar progressBar;
 	private JTable progressDataTable = new JTable();
 	private DefaultTableModel progressModel;
-	private JComboBox<Object> bookShelfCombo;
-	DefaultComboBoxModel<Object> comboModel;
+	public JComboBox<String> bookShelfCombo;
+	DefaultComboBoxModel<String> comboModel;
 	private JTextField inputTodayPages;
 	private JScrollPane scrollPane;
 	private static int todayProgress;
 	ActionList actionList;
 	Controller controller;
+
 	/**
 	 * Create the frame.
 	 */
-	public TestUI(int userId, int previousBookId)  {
+	public TestUI(int userId, int previousBookId) {
 		frame = new JFrame();
 		frame.setTitle("Book MAP");
 		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		frame.setBounds(1100, 100, 788, 381);
+		frame.setBounds(1100, 100, 714, 381);
 		//
 		// contentPane = new JPanel();
 		// contentPane.setBorder(new EmptyBorder(5, 5, 5, 5));
@@ -76,26 +76,26 @@ public class TestUI  {
 		panel.setBackground(new Color(48, 48, 48));
 		this.userId = userId;
 		this.bookId = previousBookId;
-		
+
+		TestUI testUI = this;
 		showC = new ShowController(userId, bookId);
 		actionList = new ActionList(this);
 		controller = new Controller();
 		GridBagLayout gbl_panel = new GridBagLayout();
-		gbl_panel.columnWidths = new int[] { 338, 91, 39, 135, 38, 0 };
-		gbl_panel.rowHeights = new int[] { 54, 21, 83, 55, 38, 25, 43 };
-		gbl_panel.columnWeights = new double[] { 0.0, 0.0, 0.0, 0.0, 0.0, Double.MIN_VALUE };
+		gbl_panel.columnWidths = new int[] { 70, 261, 96, 84, 147, 19, 0 };
+		gbl_panel.rowHeights = new int[] { 54, 21, 83, 55, 36, 25, 43 };
+		gbl_panel.columnWeights = new double[] { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, Double.MIN_VALUE };
 		gbl_panel.rowWeights = new double[] { 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0 };
-		
-		panel.setLayout(gbl_panel);
 
-		
+		panel.setLayout(gbl_panel);
 
 		bookTitleLabel = new JLabel();
 		bookTitleLabel.setHorizontalAlignment(SwingConstants.RIGHT);
 		adjustableFontSize(userId, bookId);
 		bookTitleLabel.setForeground(new Color(51, 153, 255));
-		
+
 		GridBagConstraints gbc_bookTitleLabel = new GridBagConstraints();
+		gbc_bookTitleLabel.gridwidth = 2;
 		gbc_bookTitleLabel.weighty = 1.0;
 		gbc_bookTitleLabel.weightx = 1.0;
 		gbc_bookTitleLabel.fill = GridBagConstraints.HORIZONTAL;
@@ -112,14 +112,8 @@ public class TestUI  {
 
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				actionList.bookListButtonAction(userId);
+				actionList.bookListButtonAction(userId, testUI);
 			}
-//				if (mBooks == null) {
-//					mBooks = new ManageBooks(userId);
-//				}
-//				mBooks.run();
-//				mBooks.updateFrame(userId);
-//			}
 		});
 
 		GridBagConstraints gbc_bookListButton = new GridBagConstraints();
@@ -127,11 +121,11 @@ public class TestUI  {
 		gbc_bookListButton.weightx = 1.0;
 		gbc_bookListButton.weighty = 1.0;
 		gbc_bookListButton.insets = new Insets(0, 0, 5, 5);
-		gbc_bookListButton.gridx = 2;
+		gbc_bookListButton.gridx = 3;
 		gbc_bookListButton.gridy = 0;
 		panel.add(bookListButton, gbc_bookListButton);
 
-		comboModel = controller.setBookList(userId, bookId);
+		comboModel = controller.setBookList(userId);
 		bookShelfCombo = new JComboBox<>(comboModel);
 		bookShelfCombo.setForeground(new Color(51, 153, 255));
 		bookShelfCombo.setBackground(new Color(48, 48, 48));
@@ -141,8 +135,7 @@ public class TestUI  {
 
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				String bookTitle;
-				bookTitle = String.valueOf(bookShelfCombo.getSelectedItem());
+				String bookTitle = String.valueOf(bookShelfCombo.getSelectedItem());
 				bookId = showC.getBookId(userId, bookTitle);
 				updateText(userId, bookId);
 			}
@@ -152,9 +145,31 @@ public class TestUI  {
 		gbc_bookShelfCombo.weighty = 1.0;
 		gbc_bookShelfCombo.weightx = 1.0;
 		gbc_bookShelfCombo.insets = new Insets(0, 10, 5, 5);
-		gbc_bookShelfCombo.gridx = 3;
+		gbc_bookShelfCombo.gridx = 4;
 		gbc_bookShelfCombo.gridy = 0;
 		panel.add(bookShelfCombo, gbc_bookShelfCombo);
+
+		changeUI = new JButton("UI");
+		changeUI.setForeground(new Color(51, 153, 255));
+		changeUI.setBackground(new Color(48, 48, 48));
+		changeUI.addActionListener(new ActionListener() {
+
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				stop();
+				Classic classic = new Classic(userId, bookId);
+				classic.run();
+			}
+
+		});
+		GridBagConstraints gbc_changeUI = new GridBagConstraints();
+		gbc_changeUI.anchor = GridBagConstraints.EAST;
+		gbc_changeUI.weighty = 1.0;
+		gbc_changeUI.weightx = 1.0;
+		gbc_changeUI.insets = new Insets(0, 10, 5, 5);
+		gbc_changeUI.gridx = 3;
+		gbc_changeUI.gridy = 1;
+		panel.add(changeUI, gbc_changeUI);
 
 		rPAnsLabel = new JLabel(controller.setRemainPageLabel(userId, bookId));
 		rPAnsLabel.setHorizontalAlignment(SwingConstants.RIGHT);
@@ -165,13 +180,12 @@ public class TestUI  {
 		gbc_rPAnsLabel.weightx = 1.0;
 		gbc_rPAnsLabel.anchor = GridBagConstraints.NORTHEAST;
 		gbc_rPAnsLabel.insets = new Insets(0, 0, 5, 5);
-		gbc_rPAnsLabel.gridx = 0;
+		gbc_rPAnsLabel.gridx = 1;
 		gbc_rPAnsLabel.gridy = 2;
 		panel.add(rPAnsLabel, gbc_rPAnsLabel);
 
-		
 		getP.add(panel, BorderLayout.CENTER);
-		
+
 		progressModel = new DefaultTableModel();
 		progressModel.addColumn("ページ数");
 		progressModel.addColumn("日時");
@@ -190,11 +204,10 @@ public class TestUI  {
 		gbc_scrollPane_1.weightx = 1.0;
 		gbc_scrollPane_1.gridheight = 2;
 		gbc_scrollPane_1.insets = new Insets(0, 10, 5, 5);
-		gbc_scrollPane_1.gridx = 1;
+		gbc_scrollPane_1.gridx = 2;
 		gbc_scrollPane_1.gridy = 2;
 		panel.add(scrollPane, gbc_scrollPane_1);
 
-		
 		progressLabel = new JLabel(controller.setProgressLabelString(userId, bookId));
 		progressLabel.setFont(new Font("MS UI Gothic", Font.PLAIN, 26));
 		progressLabel.setForeground(new Color(51, 153, 255));
@@ -203,7 +216,7 @@ public class TestUI  {
 		gbc_progressLabel.weightx = 1.0;
 		gbc_progressLabel.gridheight = 1;
 		gbc_progressLabel.insets = new Insets(0, 10, 5, 5);
-		gbc_progressLabel.gridx = 3;
+		gbc_progressLabel.gridx = 4;
 		gbc_progressLabel.gridy = 1;
 		panel.add(progressLabel, gbc_progressLabel);
 
@@ -218,7 +231,7 @@ public class TestUI  {
 		gbc_progressBar.fill = GridBagConstraints.BOTH;
 		gbc_progressBar.gridheight = 4;
 		gbc_progressBar.insets = new Insets(0, 10, 5, 5);
-		gbc_progressBar.gridx = 3;
+		gbc_progressBar.gridx = 4;
 		gbc_progressBar.gridy = 2;
 		panel.add(progressBar, gbc_progressBar);
 
@@ -230,7 +243,7 @@ public class TestUI  {
 		gbc_avgPAnsLabel.weightx = 1.0;
 		gbc_avgPAnsLabel.fill = GridBagConstraints.HORIZONTAL;
 		gbc_avgPAnsLabel.insets = new Insets(0, 0, 5, 5);
-		gbc_avgPAnsLabel.gridx = 0;
+		gbc_avgPAnsLabel.gridx = 1;
 		gbc_avgPAnsLabel.gridy = 4;
 		panel.add(avgPAnsLabel, gbc_avgPAnsLabel);
 
@@ -244,7 +257,7 @@ public class TestUI  {
 		gbc_inputTodayPages.weighty = 1.0;
 		gbc_inputTodayPages.weightx = 1.0;
 		gbc_inputTodayPages.insets = new Insets(0, 10, 5, 5);
-		gbc_inputTodayPages.gridx = 1;
+		gbc_inputTodayPages.gridx = 2;
 		gbc_inputTodayPages.gridy = 4;
 		panel.add(inputTodayPages, gbc_inputTodayPages);
 
@@ -253,24 +266,24 @@ public class TestUI  {
 		inputButton.setBackground(new Color(48, 48, 48));
 		inputButton.addActionListener(new ActionListener() {
 
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                if (inputTodayPages.getText() == null) {
-                    return;
-                } else {
-                    todayProgress = Integer.valueOf(inputTodayPages.getText());
-                    showC.addRecentData(userId, bookId, todayProgress);
-                }
-                updateText(userId, bookId);
-            }
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				if (inputTodayPages.getText() == null) {
+					return;
+				} else {
+					todayProgress = Integer.valueOf(inputTodayPages.getText());
+					showC.addRecentData(userId, bookId, todayProgress);
+				}
+				updateText(userId, bookId);
+			}
 
-        });
+		});
 		GridBagConstraints gbc_inputButton = new GridBagConstraints();
 		gbc_inputButton.anchor = GridBagConstraints.EAST;
 		gbc_inputButton.weighty = 1.0;
 		gbc_inputButton.weightx = 1.0;
 		gbc_inputButton.insets = new Insets(0, 0, 5, 5);
-		gbc_inputButton.gridx = 2;
+		gbc_inputButton.gridx = 3;
 		gbc_inputButton.gridy = 4;
 		panel.add(inputButton, gbc_inputButton);
 
@@ -282,7 +295,7 @@ public class TestUI  {
 		gbc_sumDaysAnsLabel.weightx = 1.0;
 		gbc_sumDaysAnsLabel.fill = GridBagConstraints.BOTH;
 		gbc_sumDaysAnsLabel.insets = new Insets(0, 0, 5, 5);
-		gbc_sumDaysAnsLabel.gridx = 0;
+		gbc_sumDaysAnsLabel.gridx = 1;
 		gbc_sumDaysAnsLabel.gridy = 5;
 		panel.add(sumDaysAnsLabel, gbc_sumDaysAnsLabel);
 
@@ -291,42 +304,20 @@ public class TestUI  {
 		deleteButton.setForeground(new Color(51, 153, 255));
 		deleteButton.addActionListener(new ActionListener() {
 
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                showC.deleteRecentData(userId, bookId);
-                updateText(userId, bookId);
-            }
-        });
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				showC.deleteRecentData(userId, bookId);
+				updateText(userId, bookId);
+			}
+		});
 		GridBagConstraints gbc_deleteButton = new GridBagConstraints();
 		gbc_deleteButton.anchor = GridBagConstraints.EAST;
 		gbc_deleteButton.weighty = 1.0;
 		gbc_deleteButton.weightx = 1.0;
 		gbc_deleteButton.insets = new Insets(0, 0, 5, 5);
-		gbc_deleteButton.gridx = 2;
+		gbc_deleteButton.gridx = 3;
 		gbc_deleteButton.gridy = 5;
 		panel.add(deleteButton, gbc_deleteButton);
-
-		changeUI = new JButton("UI");
-		changeUI.setForeground(new Color(51, 153, 255));
-		changeUI.setBackground(new Color(48, 48, 48));
-		changeUI.addActionListener(new ActionListener() {
-
-            @Override
-            public void actionPerformed(ActionEvent e) {
-				stop();
-                Window window = new Window(userId, bookId);
-				window.run();
-            }
-
-        });
-		GridBagConstraints gbc_changeUI = new GridBagConstraints();
-		gbc_changeUI.anchor = GridBagConstraints.WEST;
-		gbc_changeUI.weighty = 1.0;
-		gbc_changeUI.weightx = 1.0;
-		gbc_changeUI.insets = new Insets(0, 10, 5, 5);
-		gbc_changeUI.gridx = 1;
-		gbc_changeUI.gridy = 5;
-		panel.add(changeUI, gbc_changeUI);
 	}
 
 	public void run() {
@@ -336,8 +327,9 @@ public class TestUI  {
 	public void stop() {
 		this.frame.setVisible(false);
 	}
+
 	public void adjustableFontSize(int userId, int bookId) {
-        String title =  controller.setBookTitle(userId, bookId);
+		String title = controller.setBookTitle(userId, bookId);
 		bookTitleLabel.setText(title);
 		int variable = title.length() / 10;
 		if (variable == 0) {
@@ -351,9 +343,18 @@ public class TestUI  {
 	public void updateText(int userId, int bookId) {
 		adjustableFontSize(userId, bookId);
 		rPAnsLabel.setText(controller.setRemainPageLabel(userId, bookId));
-		avgPAnsLabel.setText(controller.setAvgPagesLabel(userId, bookId) + "P");
+		avgPAnsLabel.setText(controller.setAvgPagesLabel(userId, bookId));
 		sumDaysAnsLabel.setText(controller.sumDays(userId, bookId) + "日");
-		comboModel = controller.setBookList(userId, bookId);
-		progressModel = controller.reloadProgressModel(progressModel, userId,bookId);
+		progressLabel.setText(controller.setProgressLabelString(userId, bookId));
+		progressModel = controller.reloadProgressModel(progressModel, userId, bookId);
+
+	}
+
+	public void updateBookShlefCombo() {
+		//int beforeIndex = bookShelfCombo.getSelectedIndex();	//選択していたアイテムの要素番号を取得
+		comboModel = controller.setBookList(userId);
+		bookShelfCombo.setModel(comboModel);
+		bookShelfCombo.setSelectedItem(null);
+
 	}
 }
